# apps/web/src/lib/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tenants": {
      "get": {
        "operationId": "TenantsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "legalName": {
                            "type": "string"
                          },
                          "tradeName": {
                            "nullable": true,
                            "type": "string"
                          },
                          "countryCode": {
                            "type": "string"
                          },
                          "baseCurrency": {
                            "type": "string"
                          },
                          "defaultLocale": {
                            "type": "string",
                            "enum": ["en-GB", "fr-FR"]
                          },
                          "timezone": {
                            "type": "string"
                          },
                          "organizationType": {
                            "type": "string",
                            "enum": ["ISSUER", "ASSET_MANAGER", "FUND"]
                          },
                          "status": {
                            "type": "string",
                            "enum": ["ACTIVE", "INACTIVE"]
                          },
                          "version": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          }
                        },
                        "required": [
                          "id",
                          "legalName",
                          "tradeName",
                          "countryCode",
                          "baseCurrency",
                          "defaultLocale",
                          "timezone",
                          "organizationType",
                          "status",
                          "version",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "pageSize": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "total": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": ["page", "pageSize", "total"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["data", "meta"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["tenants"]
      },
      "post": {
        "operationId": "TenantsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "legalName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "tradeName": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 200
                  },
                  "countryCode": {
                    "type": "string",
                    "pattern": "^[A-Z]{2}$"
                  },
                  "baseCurrency": {
                    "type": "string",
                    "pattern": "^[A-Z]{3}$"
                  },
                  "defaultLocale": {
                    "default": "en-GB",
                    "type": "string",
                    "enum": ["en-GB", "fr-FR"]
                  },
                  "timezone": {
                    "default": "Europe/Paris",
                    "type": "string"
                  },
                  "organizationType": {
                    "type": "string",
                    "enum": ["ISSUER", "ASSET_MANAGER", "FUND"]
                  },
                  "firstAdministrator": {
                    "type": "object",
                    "properties": {
                      "email": {
                        "type": "string",
                        "maxLength": 254,
                        "format": "email",
                        "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                      },
                      "name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 200
                      },
                      "locale": {
                        "default": "en-GB",
                        "type": "string",
                        "enum": ["en-GB", "fr-FR"]
                      }
                    },
                    "required": ["email", "name", "locale"],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "legalName",
                  "countryCode",
                  "baseCurrency",
                  "defaultLocale",
                  "timezone",
                  "organizationType",
                  "firstAdministrator"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "legalName": {
                      "type": "string"
                    },
                    "tradeName": {
                      "nullable": true,
                      "type": "string"
                    },
                    "countryCode": {
                      "type": "string"
                    },
                    "baseCurrency": {
                      "type": "string"
                    },
                    "defaultLocale": {
                      "type": "string",
                      "enum": ["en-GB", "fr-FR"]
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "organizationType": {
                      "type": "string",
                      "enum": ["ISSUER", "ASSET_MANAGER", "FUND"]
                    },
                    "status": {
                      "type": "string",
                      "enum": ["ACTIVE", "INACTIVE"]
                    },
                    "version": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    }
                  },
                  "required": [
                    "id",
                    "legalName",
                    "tradeName",
                    "countryCode",
                    "baseCurrency",
                    "defaultLocale",
                    "timezone",
                    "organizationType",
                    "status",
                    "version",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["tenants"]
      }
    },
    "/api/v1/tenants/{id}": {
      "get": {
        "operationId": "TenantsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "legalName": {
                      "type": "string"
                    },
                    "tradeName": {
                      "nullable": true,
                      "type": "string"
                    },
                    "countryCode": {
                      "type": "string"
                    },
                    "baseCurrency": {
                      "type": "string"
                    },
                    "defaultLocale": {
                      "type": "string",
                      "enum": ["en-GB", "fr-FR"]
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "organizationType": {
                      "type": "string",
                      "enum": ["ISSUER", "ASSET_MANAGER", "FUND"]
                    },
                    "status": {
                      "type": "string",
                      "enum": ["ACTIVE", "INACTIVE"]
                    },
                    "version": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    }
                  },
                  "required": [
                    "id",
                    "legalName",
                    "tradeName",
                    "countryCode",
                    "baseCurrency",
                    "defaultLocale",
                    "timezone",
                    "organizationType",
                    "status",
                    "version",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["tenants"]
      },
      "patch": {
        "operationId": "TenantsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "legalName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "tradeName": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 200
                  },
                  "defaultLocale": {
                    "type": "string",
                    "enum": ["en-GB", "fr-FR"]
                  },
                  "timezone": {
                    "type": "string"
                  },
                  "organizationType": {
                    "type": "string",
                    "enum": ["ISSUER", "ASSET_MANAGER", "FUND"]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "legalName": {
                      "type": "string"
                    },
                    "tradeName": {
                      "nullable": true,
                      "type": "string"
                    },
                    "countryCode": {
                      "type": "string"
                    },
                    "baseCurrency": {
                      "type": "string"
                    },
                    "defaultLocale": {
                      "type": "string",
                      "enum": ["en-GB", "fr-FR"]
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "organizationType": {
                      "type": "string",
                      "enum": ["ISSUER", "ASSET_MANAGER", "FUND"]
                    },
                    "status": {
                      "type": "string",
                      "enum": ["ACTIVE", "INACTIVE"]
                    },
                    "version": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    }
                  },
                  "required": [
                    "id",
                    "legalName",
                    "tradeName",
                    "countryCode",
                    "baseCurrency",
                    "defaultLocale",
                    "timezone",
                    "organizationType",
                    "status",
                    "version",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["tenants"]
      }
    },
    "/api/v1/tenants/{id}/activate": {
      "post": {
        "operationId": "TenantsController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "legalName": {
                      "type": "string"
                    },
                    "tradeName": {
                      "nullable": true,
                      "type": "string"
                    },
                    "countryCode": {
                      "type": "string"
                    },
                    "baseCurrency": {
                      "type": "string"
                    },
                    "defaultLocale": {
                      "type": "string",
                      "enum": ["en-GB", "fr-FR"]
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "organizationType": {
                      "type": "string",
                      "enum": ["ISSUER", "ASSET_MANAGER", "FUND"]
                    },
                    "status": {
                      "type": "string",
                      "enum": ["ACTIVE", "INACTIVE"]
                    },
                    "version": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    }
                  },
                  "required": [
                    "id",
                    "legalName",
                    "tradeName",
                    "countryCode",
                    "baseCurrency",
                    "defaultLocale",
                    "timezone",
                    "organizationType",
                    "status",
                    "version",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["tenants"]
      }
    },
    "/api/v1/tenants/{id}/deactivate": {
      "post": {
        "operationId": "TenantsController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "legalName": {
                      "type": "string"
                    },
                    "tradeName": {
                      "nullable": true,
                      "type": "string"
                    },
                    "countryCode": {
                      "type": "string"
                    },
                    "baseCurrency": {
                      "type": "string"
                    },
                    "defaultLocale": {
                      "type": "string",
                      "enum": ["en-GB", "fr-FR"]
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "organizationType": {
                      "type": "string",
                      "enum": ["ISSUER", "ASSET_MANAGER", "FUND"]
                    },
                    "status": {
                      "type": "string",
                      "enum": ["ACTIVE", "INACTIVE"]
                    },
                    "version": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    }
                  },
                  "required": [
                    "id",
                    "legalName",
                    "tradeName",
                    "countryCode",
                    "baseCurrency",
                    "defaultLocale",
                    "timezone",
                    "organizationType",
                    "status",
                    "version",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["tenants"]
      }
    },
    "/api/v1/tenants/{id}/administrators": {
      "post": {
        "operationId": "TenantsController_inviteAdministrator",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "maxLength": 254,
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "locale": {
                    "default": "en-GB",
                    "type": "string",
                    "enum": ["en-GB", "fr-FR"]
                  }
                },
                "required": ["email", "name", "locale"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    }
                  },
                  "required": ["id", "expiresAt"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["tenants"]
      }
    },
...
  },
  "info": {
    "title": "Virtus Assets API",
    "description": "Demonstration environment — fictitious data.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
...
}